extract SolutionValidator from SessionLogFormatter
