+++ src/main/java/org/cicirello/ibp/SessionLogFormatter.java
@@ -28,9 +28,6 @@ import java.time.Duration;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Formatter;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Random;
 
 /** Formats SessionLogs. */
 final class SessionLogFormatter {
@@ -39,11 +36,13 @@ final class SessionLogFormatter {
   private final RecordList records;
   private final int[] successfulMoves;
   private final int[] failedMoves;
+  private final SolutionValidator validator;
 
   SessionLogFormatter(RecordList records, int[] successfulMoves, int[] failedMoves) {
     this.records = records;
     this.successfulMoves = successfulMoves;
     this.failedMoves = failedMoves;
+    validator = new SolutionValidator();
     ArrayList<String> alertList = new ArrayList<String>();
     String summary = formatSummaryStats();
     String allActions = formatAllLoggedActions(alertList);
@@ -101,7 +100,7 @@ final class SessionLogFormatter {
           log = records.get(i);
           String solutionData = log.getData();
           int m =
-              validateSolution(
+              validator.validateSolution(
                   completedData,
                   solutionData,
                   currentModeName,
@@ -139,266 +138,6 @@ final class SessionLogFormatter {
       s.append("</table>");
     }
     return s.toString();
-  }
-
-  int validateSolution(
-      String completedData,
-      String solutionData,
-      String currentModeName,
-      String currentInstance,
-      ArrayList<String> completionTableRows,
-      ArrayList<String> alertList) {
-    int modeNum = extractModeNum(completedData);
-    String instance = extractInstance(completedData);
-    boolean proceed = true;
-    if (!instance.equals(currentInstance)) {
-      alertList.add("Solution encoded doesn't correspond to instance.");
-      proceed = false;
-    }
-    String modeName = extractModeName(completedData);
-    if (!modeName.equals(currentModeName)) {
-      alertList.add("Solution encoded doesn't correspond to mode.");
-      proceed = false;
-    }
-    if (!modeName.equals(ApplicationState.modeIntToModeName(modeNum))) {
-      alertList.add("Mode name and internal mode number are inconsistent.");
-      proceed = false;
-    }
-    if (modeNum == 0) {
-      alertList.add(
-          "Application doesn't log solutions in practice mode, but session log contains a practice mode solution.");
-      proceed = false;
-    }
-    if (proceed) {
-      int i = solutionData.indexOf("ItemSequence=");
-      int j = solutionData.indexOf(",", 13);
-      int x = solutionData.indexOf("BinSequence=");
-      int k = x + 12;
-      if (i < 0 || j < 0 || x < 0 || k >= solutionData.length()) {
-        alertList.add("Encoded solution is malformed.");
-      } else {
-        String[] strBins = solutionData.substring(k).split(" ");
-        int[] bins = new int[strBins.length];
-        try {
-          for (int b = 0; b < strBins.length; b++) {
-            bins[b] = Integer.parseInt(strBins[b]);
-          }
-        } catch (NumberFormatException ex) {
-          alertList.add("Expected integer bin numbers in solution.");
-          proceed = false;
-        }
-        String[] pairs = solutionData.substring(i + 13, j).split(" ");
-        if (pairs.length != 40) {
-          alertList.add("Wrong number of items in solution");
-          proceed = false;
-        }
-        int[] sizes = new int[20];
-        String[] items = new String[20];
-        try {
-          for (int b = 0, index = 0; b < pairs.length; b += 2, index++) {
-            sizes[index] = Integer.parseInt(pairs[b + 1]);
-            items[index] = pairs[b];
-          }
-        } catch (NumberFormatException ex) {
-          alertList.add("Expected integer item sizes in solution.");
-          proceed = false;
-        }
-        if (proceed) {
-          if (checkInstance(sizes, items, instance, alertList)
-              && checkItemOrder(sizes, items, modeNum, alertList)
-              && checkBins(sizes, bins, modeNum, alertList)) {
-            int numBins = binCount(bins);
-            completionTableRows.add(
-                "<tr>\n<td style=\"text-align:left\">"
-                    + instance
-                    + "</td>\n<td style=\"text-align:left\">"
-                    + modeName
-                    + "</td>\n<td style=\"text-align:left\">Valid solution with "
-                    + numBins
-                    + " bins</td>\n</tr>\n");
-            return modeNum;
-          }
-        }
-      }
-    }
-    return -1;
-  }
-
-  int binCount(int[] bins) {
-    HashSet<Integer> unique = new HashSet<Integer>();
-    for (int b : bins) {
-      unique.add(b);
-    }
-    return unique.size();
-  }
-
-  boolean checkBins(int[] sizes, int[] bins, int modeNum, ArrayList<String> alertList) {
-    int[] binCapacities = new int[sizes.length];
-    for (int i = 0; i < binCapacities.length; i++) {
-      binCapacities[i] = 100;
-    }
-    if (modeNum == ApplicationState.MODE_FIRST_FIT
-        || modeNum == ApplicationState.MODE_FIRST_FIT_DECREASING) {
-      for (int i = 0; i < sizes.length; i++) {
-        boolean foundIt = false;
-        for (int j = 0; !foundIt; j++) {
-          if (sizes[i] <= binCapacities[j]) {
-            if (j + 1 == bins[i]) {
-              binCapacities[j] -= sizes[i];
-              foundIt = true;
-            } else {
-              alertList.add("Items put in incorrect bins for chosen mode.");
-              return false;
-            }
-          }
-        }
-      }
-    } else if (modeNum == ApplicationState.MODE_BEST_FIT
-        || modeNum == ApplicationState.MODE_BEST_FIT_DECREASING) {
-      for (int i = 0; i < sizes.length; i++) {
-        for (int j = 0; j < binCapacities.length; j++) {
-          if (j + 1 != bins[i]
-              && sizes[i] <= binCapacities[j]
-              && binCapacities[j] < binCapacities[bins[i] - 1]) {
-            alertList.add("Items put in incorrect bins for chosen mode.");
-            return false;
-          }
-        }
-        binCapacities[bins[i] - 1] -= sizes[i];
-      }
-    } else {
-      return false;
-    }
-    return true;
-  }
-
-  boolean checkItemOrder(int[] sizes, String[] items, int modeNum, ArrayList<String> alertList) {
-    if (modeNum == ApplicationState.MODE_FIRST_FIT || modeNum == ApplicationState.MODE_BEST_FIT) {
-      char c = 'A';
-      for (int i = 0; i < items.length; i++, c = (char) (c + 1)) {
-        if (!items[i].equals("" + c)) {
-          alertList.add("Items used in incorrect order for chosen mode.");
-          return false;
-        }
-      }
-    } else if (modeNum == ApplicationState.MODE_FIRST_FIT_DECREASING
-        || modeNum == ApplicationState.MODE_BEST_FIT_DECREASING) {
-      for (int i = 1; i < sizes.length; i++) {
-        if (sizes[i] > sizes[i - 1]) {
-          alertList.add("Items used in incorrect order for chosen mode.");
-          return false;
-        }
-      }
-    } else {
-      return false;
-    }
-    return true;
-  }
-
-  boolean checkInstance(int[] sizes, String[] items, String instance, ArrayList<String> alertList) {
-    boolean goodInstance = true;
-    HashMap<String, Integer> foundItems = new HashMap<String, Integer>();
-    for (int i = 0; i < sizes.length; i++) {
-      if (foundItems.containsKey(items[i])) {
-        alertList.add("Duplicate items in solution.");
-        goodInstance = false;
-      } else {
-        foundItems.put(items[i], sizes[i]);
-      }
-    }
-    if (instance.equals("Default")) {
-      int[] weights = {36, 33, 39, 43, 7, 19, 37, 8, 29, 28, 37, 23, 29, 10, 22, 11, 33, 9, 17, 30};
-      char c = 'A';
-      for (int i = 0; i < weights.length; i++, c = (char) (c + 1)) {
-        if (foundItems.containsKey("" + c)) {
-          int w = foundItems.get("" + c);
-          if (w != weights[i]) {
-            alertList.add("Wrong item size found in solution of default instance.");
-            goodInstance = false;
-          }
-        } else {
-          alertList.add("Unknown items found in solution of default instance.");
-          goodInstance = false;
-        }
-      }
-    } else if (instance.equals("Random")) {
-      char c = 'A';
-      for (int i = 0; i < items.length; i++, c = (char) (c + 1)) {
-        if (foundItems.containsKey("" + c)) {
-          int w = foundItems.get("" + c);
-          if (w < 20 || w > 50) {
-            alertList.add("Wrong item size found in solution of random instance.");
-            goodInstance = false;
-          }
-        } else {
-          alertList.add("Unknown items found in solution of random instance.");
-          goodInstance = false;
-        }
-      }
-    } else if (instance.startsWith("#")) {
-      try {
-        long seed = Long.parseLong(instance.substring(1));
-        int[] weights = ApplicationState.createRandomItemSizes(20, 50, 20, new Random(seed));
-        char c = 'A';
-        for (int i = 0; i < weights.length; i++, c = (char) (c + 1)) {
-          if (foundItems.containsKey("" + c)) {
-            int w = foundItems.get("" + c);
-            if (w != weights[i]) {
-              alertList.add("Wrong item size found in solution of instance: " + instance);
-              goodInstance = false;
-            }
-          } else {
-            alertList.add("Unknown items found in solution of instance: " + instance);
-            goodInstance = false;
-          }
-        }
-      } catch (NumberFormatException ex) {
-        alertList.add("Malformed instance number in solution.");
-        goodInstance = false;
-      }
-    } else {
-      alertList.add("Unknown instance type found in solution.");
-      goodInstance = false;
-    }
-    return goodInstance;
-  }
-
-  int extractModeNum(String completedData) {
-    int i = completedData.indexOf("ModeNum=");
-    if (i >= 0) {
-      i += 8;
-      int j = completedData.indexOf(",", i);
-      if (j >= 0) {
-        String s = completedData.substring(i, j);
-        try {
-          int num = Integer.parseInt(s);
-          return num;
-        } catch (NumberFormatException ex) {
-          return -1;
-        }
-      }
-    }
-    return -1;
-  }
-
-  String extractModeName(String completedData) {
-    int i = completedData.indexOf("Mode=");
-    if (i >= 0) {
-      return completedData.substring(i + 5);
-    }
-    return "";
-  }
-
-  String extractInstance(String completedData) {
-    int i = completedData.indexOf("Instance=");
-    if (i >= 0) {
-      i += 9;
-      int j = completedData.indexOf(",", i);
-      if (j >= 0) {
-        return completedData.substring(i, j);
-      }
-    }
-    return "";
   }
 
   /**
@@ -484,15 +223,15 @@ final class SessionLogFormatter {
   }
 
   String formatCompletedData(String data, ArrayList<String> alertList) {
-    int modeNum = extractModeNum(data);
+    int modeNum = validator.extractModeNum(data);
     if (modeNum < 0) {
       return malformed(alertList);
     }
-    String instance = extractInstance(data);
+    String instance = validator.extractInstance(data);
     if (instance.length() == 0) {
       return malformed(alertList);
     }
-    String mode = extractModeName(data);
+    String mode = validator.extractModeName(data);
     if (mode.length() == 0) {
       return malformed(alertList);
     }
